+++ src/app/services/clipserver-connection/clipserver-connection.service.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@angular/core';
-import { WSServerStatus,GlobalConstants } from "../../global-constants";
+import { WSServerStatus,GlobalConstants } from "../../shared/config/global-constants";
 import { Observable, Observer } from 'rxjs';
 import { AnonymousSubject } from 'rxjs/internal/Subject';
 import { Subject } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { GlobalConstantsService } from '../config/global-constants.service';
+import { GlobalConstantsService } from '../../shared/config/services/global-constants.service';
 
 //const URL = GlobalConstants.clipServerURL;
 let statusConnected = {'wsstatus':'connected'};
