+++ src/app/shared/config/global-constants.ts
