+++ src/app/shared/config/local-config-example.ts
