+++ src/app/shared/config/services/global-constants.service.ts
