+++ src/app/shared/interfaces/gui-action.ts
