+++ src/app/shared/interfaces/vbs-task-interface.ts
