+++ src/app/shared/safe.pipe.ts
